feat: move UserProfile conversations and new threads feed filters + VMs to commons
Moves UserProfileConversationsFeedFilter, UserProfileNewThreadFeedFilter
and their ViewModels to commons. These use IAccount + ICacheProvider
abstractions, making them available for iOS/desktop targets.

Android wrappers delegate to commons and add ViewModelProvider.Factory.
Part of KMP iOS migration tracked in #2238.
