+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/conversations/dal/UserProfileConversationsFeedFilter.kt
@@ -22,63 +22,5 @@
 
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.profile.conversations.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
-import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
-import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
-import com.vitorpamplona.quartz.nip35Torrents.TorrentCommentEvent
-import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
-import com.vitorpamplona.quartz.nip88Polls.response.PollResponseEvent
-import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceReplyEvent
-import com.vitorpamplona.quartz.nipA4PublicMessages.PublicMessageEvent
-
-class UserProfileConversationsFeedFilter(
-    val user: User,
-    val account: Account,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + user.pubkeyHex
-
-    override fun feed(): List<Note> {
-        val notes =
-            LocalCache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it)
-            }
-
-        val longFormNotes =
-            LocalCache.addressables.filterIntoSet { _, it ->
-                acceptableEvent(it)
-            }
-
-        return sort(notes + longFormNotes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> = collection.filterTo(HashSet()) { acceptableEvent(it) }
-
-    fun acceptableEvent(it: Note): Boolean =
-        it.author == user &&
-            (
-                it.event is TextNoteEvent ||
-                    it.event is ZapPollEvent ||
-                    it.event is PollResponseEvent ||
-                    it.event is ChannelMessageEvent ||
-                    it.event is LiveActivitiesChatMessageEvent ||
-                    it.event is CommentEvent ||
-                    it.event is VoiceReplyEvent ||
-                    it.event is PublicMessageEvent ||
-                    it.event is TorrentCommentEvent
-            ) &&
-            !it.isNewThread() &&
-            account.isAcceptable(it)
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-
-    override fun limit() = 200
-}
+// Re-export from commons for backwards compatibility
+typealias UserProfileConversationsFeedFilter = com.vitorpamplona.amethyst.commons.viewmodels.profile.UserProfileConversationsFeedFilter
