+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/profile/newthreads/dal/UserProfileNewThreadsFeedViewModel.kt
@@ -24,14 +24,18 @@ import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.vitorpamplona.amethyst.model.Account
+import com.vitorpamplona.amethyst.model.LocalCache
 import com.vitorpamplona.amethyst.model.User
-import com.vitorpamplona.amethyst.ui.screen.AndroidFeedViewModel
 
 @Stable
 class UserProfileNewThreadsFeedViewModel(
-    val user: User,
-    val account: Account,
-) : AndroidFeedViewModel(UserProfileNewThreadFeedFilter(user, account)) {
+    user: User,
+    account: Account,
+) : com.vitorpamplona.amethyst.commons.viewmodels.profile.UserProfileNewThreadsFeedViewModel(
+        user,
+        account,
+        LocalCache,
+    ) {
     class Factory(
         val user: User,
         val account: Account,
