+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/viewmodels/profile/UserProfileConversationsFeedFilter.kt
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+@file:Suppress("DEPRECATION")
+
+package com.vitorpamplona.amethyst.commons.viewmodels.profile
+
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.Note
+import com.vitorpamplona.amethyst.commons.model.User
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.ui.feeds.AdditiveFeedFilter
+import com.vitorpamplona.amethyst.commons.ui.feeds.DefaultFeedOrder
+import com.vitorpamplona.quartz.experimental.zapPolls.ZapPollEvent
+import com.vitorpamplona.quartz.nip10Notes.TextNoteEvent
+import com.vitorpamplona.quartz.nip22Comments.CommentEvent
+import com.vitorpamplona.quartz.nip28PublicChat.message.ChannelMessageEvent
+import com.vitorpamplona.quartz.nip35Torrents.TorrentCommentEvent
+import com.vitorpamplona.quartz.nip53LiveActivities.chat.LiveActivitiesChatMessageEvent
+import com.vitorpamplona.quartz.nip88Polls.response.PollResponseEvent
+import com.vitorpamplona.quartz.nipA0VoiceMessages.VoiceReplyEvent
+import com.vitorpamplona.quartz.nipA4PublicMessages.PublicMessageEvent
+
+class UserProfileConversationsFeedFilter(
+    val user: User,
+    val account: IAccount,
+    val cacheProvider: ICacheProvider,
+) : AdditiveFeedFilter<Note>() {
+    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + user.pubkeyHex
+
+    override fun feed(): List<Note> {
+        val notes =
+            cacheProvider.filterNotes { _, it ->
+                acceptableEvent(it)
+            }
+
+        val longFormNotes =
+            cacheProvider.filterAddressableNotes { it ->
+                acceptableEvent(it)
+            }
+
+        return sort(notes + longFormNotes)
+    }
+
+    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
+
+    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> = collection.filterTo(HashSet()) { acceptableEvent(it) }
+
+    fun acceptableEvent(it: Note): Boolean =
+        it.author == user &&
+            (
+                it.event is TextNoteEvent ||
+                    it.event is ZapPollEvent ||
+                    it.event is PollResponseEvent ||
+                    it.event is ChannelMessageEvent ||
+                    it.event is LiveActivitiesChatMessageEvent ||
+                    it.event is CommentEvent ||
+                    it.event is VoiceReplyEvent ||
+                    it.event is PublicMessageEvent ||
+                    it.event is TorrentCommentEvent
+            ) &&
+            !it.isNewThread() &&
+            account.isAcceptable(it)
+
+    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
+
+    override fun limit() = 200
+}
